only delete if needed

+++ src/mavros_offboard/src/run_traj.py
@@ -407,7 +407,8 @@ class MavrosOffboardSuctionMission():
         self.ros_log_info("Waiting Over")
     
     def confirm_next_stage(self, message):
-        rospy.delete_param('mission_confirm')
+        if rospy.has_param('mission_confirm'):
+            rospy.delete_param('mission_confirm')
         rospy.loginfo(message + " (set ROS parameter 'mission_confirm' to 'confirm' or 'stop')")
         while not rospy.is_shutdown():
             if rospy.has_param('mission_confirm'):
